+++ conjur/list/list_controller.py
@@ -6,11 +6,9 @@ ListController module
 This module is the controller that facilitates all list actions
 required to successfully execute the LIST command
 """
+# Builtins
 import json
-import logging
 import sys
-
-from Utils.utils import Utils
 
 
 class ListController:
@@ -22,9 +20,6 @@ class ListController:
 
     def __init__(self, ssl_verify, list_logic, list_data):
         self.ssl_verify = ssl_verify
-        if self.ssl_verify is False:
-            Utils.get_insecure_warning()
-
         self.list_logic = list_logic
         self.list_data = list_data
 
@@ -32,12 +27,7 @@ class ListController:
         """
         Method that facilitates all method calls in this class
         """
-        list_constraints = self.list_data.list_dictify()
-        # pylint: disable=logging-fstring-interpolation
-        logging.debug(f"Executing list command with the following "\
-                      f"constraints: {list_constraints}")
-
-        result = self.list_logic.list(list_constraints)
+        result = self.list_logic.list(self.list_data)
         self.print_json_result(result)
 
     @classmethod