+++ conjur/list/list_data.py
@@ -12,14 +12,13 @@ class ListData:
     """
     Used for organizing the the params the user passed in to execute the list command
     """
-    # pylint: disable=line-too-long,too-many-arguments
-    def __init__(self, kind=None, inspect=False, search=None, limit=None, offset=None, acting_as=None):
-        self.kind = kind
-        self.inspect = inspect
-        self.search = search
-        self.limit = limit
-        self.offset = offset
-        self.acting_as = acting_as
+    def __init__(self, **argParams):
+        self.kind = argParams['kind']
+        self.inspect = argParams['inspect']
+        self.search = argParams['search']
+        self.limit = argParams['limit']
+        self.offset = argParams['offset']
+        self.acting_as = argParams['acting_as']
 
     def list_dictify(self):
         """
@@ -27,16 +26,13 @@ class ListData:
         """
         list_dict={}
         for attr, value in self.__dict__.items():
-            if value is not None and value is not False:
+            if value:
                 list_dict[str(attr)] = value
 
         return list_dict
 
     def __repr__(self):
         # pylint: disable=line-too-long
-        return "{ 'kind': %r, 'inspect': %r, 'search': %r, 'limit': %r, 'offset': %r, 'role': %r }" % (self.kind,
-                                                                                                       self.inspect,
-                                                                                                       self.search,
-                                                                                                       self.limit,
-                                                                                                       self.offset,
-                                                                                                       self.acting_as)
+        return f"{{ 'kind': {self.kind}, 'inspect': {self.inspect}, " \
+               f"'search': {self.search}, 'limit': {self.limit}, " \
+               f"'offset': {self.offset}, 'role': {self.acting_as} }}"